feat: Include dataset keys from ExtendedLoader into Loader
