+++ src/atc/utils/MockLoader.py
@@ -1,4 +1,4 @@
-from typing import Dict
+from typing import Dict, List
 from unittest.mock import MagicMock
 
 from pyspark.sql import DataFrame
@@ -8,8 +8,20 @@ from atc.etl.types import dataset_group
 
 
 class MockLoader(Loader, MagicMock):
-    def __init__(self, *args, **kwargs):
-        MagicMock.__init__(self, *args, **kwargs)
+    def __init__(
+        self,
+        dataset_input_key: str = None,
+        dataset_input_key_list: List[str] = None,
+        *args,
+        **kwargs
+    ):
+        MagicMock.__init__(
+            self,
+            dataset_input_key=dataset_input_key,
+            dataset_input_key_list=dataset_input_key_list,
+            *args,
+            **kwargs
+        )
         self.saved: Dict[str, DataFrame] = {}
 
     def save(self, df: DataFrame) -> None:
